Refactoring loot generation code to make it work a bit better. Adding LootEnums header to store all loot related enums
