+++ RobotGladiator/Source/RobotGladiator/Private/LootEnums.cpp
@@ -1,0 +1,5 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "LootEnums.h"
+
