+++ RobotGladiator/Source/RobotGladiator/Public/Armor.h
@@ -4,16 +4,9 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/Actor.h"
+#include "LootEnums.h"
 #include "Armor.generated.h"
 
-UENUM(BlueprintType)
-enum EArmorTypes
-{
-	HEAD		UMETA(DisplayName = "Head"),
-	CHEST		UMETA(DisplayName = "Chest"),
-	ARMS		UMETA(DisplayName = "Arms"),
-	LEGS		UMETA(DisplayName = "Legs"),
-};
 
 UCLASS()
 class ROBOTGLADIATOR_API AArmor : public AActor
@@ -23,6 +16,9 @@ class ROBOTGLADIATOR_API AArmor : public AActor
 public:	
 	// Sets default values for this actor's properties
 	AArmor();
+
+	UFUNCTION(BlueprintCallable)
+	void SetArmorStats(int defense, ERarities rarity, EArmorTypes type);
 
 protected:
 	// Called when the game starts or when spawned
@@ -37,4 +33,10 @@ public:
 	UStaticMeshComponent* Mesh;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (ExposeOnSpawn="true"))
 	FName ArmorName;
+	UPROPERTY(EditAnywhere, BlueprintReadWrite)
+	TEnumAsByte<ERarities> Rarity;
+	UPROPERTY(EditAnywhere, BlueprintReadWrite)
+	TEnumAsByte<EArmorTypes> ArmorType;
+	UPROPERTY(EditAnywhere, BlueprintReadWrite)
+	int DefenseStat;
 };
